requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python-headless>=4.8.0
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0
psycopg2-binary>=2.9.9
